ItemDatabase 스크립트 작성, 아이템 데이터 (소비 / 장비) 로드 테스트
BGdatabase 로드 테스트 및 Item_Equip 테이블 추가, 데이터타입미스 수정

+++ ReBloom/Assets/Scripts/Item/2. Core/ItemTypes/ConsumableItemData.cs
@@ -28,7 +28,7 @@ public class ConsumableItemData : ItemBase
     private BGField<float> HP;
     private BGField<float> fieldHP;
     private BGField<float> Temp;
-    private BGField<float> Range;
+    private BGField<int> Range;
     private BGField<float> Duration;
     private BGField<string> ImgPath;
     private BGField<string> Description;
@@ -57,7 +57,7 @@ public class ConsumableItemData : ItemBase
         HP = meta.GetField<float>("Hunger");
         fieldHP = meta.GetField<float>("HP");
         Temp = meta.GetField<float>("Temp");
-        Range = meta.GetField<float>("Range");
+        Range = meta.GetField<int>("Range");
         Duration = meta.GetField<float>("Duration");
         ImgPath = meta.GetField<string>("ImgPath");
         Description = meta.GetField<string>("Description");
